Integrated last.fm history in the history page in the UI
