+++ app/api/library.py
@@ -695,80 +695,6 @@ async def get_recently_added_albums(
     return results
 
 
-@router.get("/api/home/recently-played-albums")
-async def get_recently_played_albums(
-    limit: int = 20, db: asyncpg.Connection = Depends(get_db)
-):
-    # Join playback_history to get specific albums
-    query = """
-        SELECT 
-            t.album, 
-            MAX(t.artwork_id) as artwork_id, 
-            MAX(a.sha1) as art_sha1,
-            COALESCE(MAX(t.album_artist), MAX(t.artist)) as artist_name,
-            MAX(CASE WHEN t.bit_depth > 16 OR t.sample_rate_hz > 44100 THEN 1 ELSE 0 END) as is_hires,
-            MIN(t.release_date) as year,
-            COUNT(DISTINCT t.id) as track_count,
-            SUM(t.duration_seconds) as total_duration,
-            t.release_mbid as release_mbid,
-            MAX(t.release_mbid) as mbid,
-            MAX(t.release_mbid) as mb_release_id,
-            MAX(t.release_group_mbid) as album_mbid,
-            (SELECT mbid FROM artist WHERE name = COALESCE(MAX(t.album_artist), MAX(t.artist)) LIMIT 1) as artist_mbid,
-            MAX(ph.timestamp) as last_played
-        FROM playback_history ph
-        JOIN track t ON ph.track_id = t.id
-        LEFT JOIN artwork a ON t.artwork_id = a.id
-        WHERE t.album IS NOT NULL
-        GROUP BY t.album, t.release_mbid
-        ORDER BY last_played DESC
-        LIMIT $1
-    """
-    rows = await db.fetch(query, limit)
-    results = []
-    for row in rows:
-        d = dict(row)
-        d.pop("artwork_id", None)
-        results.append(d)
-    return results
-
-
-@router.get("/api/home/recently-played-artists")
-async def get_recently_played_artists(
-    limit: int = 20, db: asyncpg.Connection = Depends(get_db)
-):
-    query = """
-        SELECT DISTINCT 
-            a.mbid,
-            a.name,
-            a.image_url, 
-            a.artwork_id,
-            ar.sha1 as art_sha1,
-            a.bio, 
-            MAX(ph.timestamp) as last_played
-        FROM playback_history ph
-        JOIN track t ON ph.track_id = t.id
-        JOIN track_artist ta ON t.id = ta.track_id
-        JOIN artist a ON ta.artist_mbid = a.mbid
-        LEFT JOIN artwork ar ON a.artwork_id = ar.id
-        WHERE a.name IS NOT NULL AND a.name != '' AND a.name != 'null'
-        GROUP BY a.mbid, a.name, a.image_url, a.artwork_id, ar.sha1, a.bio
-        ORDER BY last_played DESC
-        LIMIT $1
-    """
-    rows = await db.fetch(query, limit)
-    return [
-        {
-            "mbid": row["mbid"],
-            "name": row["name"],
-            "image_url": row["image_url"],
-            "art_sha1": sha1_to_hex(row["art_sha1"]),
-            "bio": row["bio"],
-        }
-        for row in rows
-    ]
-
-
 @router.get("/api/home/discover-artists")
 async def get_discover_artists(
     limit: int = 20, db: asyncpg.Connection = Depends(get_db)
