+++ app/api/player.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, Request, HTTPException, Header, Response
+from fastapi import APIRouter, Depends, Request, HTTPException, Header
 import os
 import asyncio
 from app.db import get_db
@@ -734,232 +734,6 @@ async def _scrobble_to_lastfm(user_id: int, track_id: int):
 @router.get("/api/client-ip")
 async def get_client_ip_endpoint(request: Request):
     return {"ip": get_client_ip(request)}
-
-
-@router.get("/api/player/history")
-async def get_playback_history(
-    response: Response,
-    scope: str = "all",
-    days: int = 7,
-    page: int = 1,
-    limit: int = 20,
-    request: Request = None,
-):
-    response.headers["Cache-Control"] = "no-cache, no-store, must-revalidate"
-    response.headers["Pragma"] = "no-cache"
-    response.headers["Expires"] = "0"
-    async for db in get_db():
-        user_row, _ = (
-            await get_session_user(db, request.cookies.get("jamarr_session"))
-            if request
-            else (None, None)
-        )
-        # Date filter
-        from datetime import timedelta
-
-        # Ensure days is valid
-        days = max(1, min(days, 365))
-        
-        where_clauses = ["DATE(h.timestamp) >= CURRENT_DATE + CAST($1 AS INTERVAL)"]
-        params_list = [timedelta(days=-(days - 1))]
-        
-        if scope == "mine" and user_row:
-            where_clauses.append("h.user_id = $2")
-            params_list.append(user_row["id"])
-            
-        where_sql = "WHERE " + " AND ".join(where_clauses)
-        
-        # Pagination
-        page = max(1, page)
-        limit = max(1, min(limit, 100))
-        offset = (page - 1) * limit
-        
-        # Add offset/limit to params (adjusting indices)
-        # params_list currently has 1 or 2 items
-        params_list.append(limit)
-        params_list.append(offset)
-        
-        # Calculate indices for SQL placeholders
-        # $1 is duration
-        # $2 is user_id (optional)
-        # Limit is next, Offset is last
-        limit_idx = len(params_list) - 1
-        offset_idx = len(params_list)
-
-        query = f"""
-            SELECT 
-                h.id, h.timestamp, h.client_ip, h.client_id, h.user_id,
-                t.id, t.title, t.artist, t.album, t.artwork_id, t.duration_seconds,
-                t.codec, t.bit_depth, t.sample_rate_hz, t.release_date,
-                t.release_mbid,
-                u.username, u.display_name, u.email,
-                a.sha1 as art_sha1
-            FROM playback_history h
-            JOIN track t ON h.track_id = t.id
-            LEFT JOIN artwork a ON t.artwork_id = a.id
-            LEFT JOIN "user" u ON u.id = h.user_id
-            {where_sql}
-            ORDER BY h.timestamp DESC
-            LIMIT ${limit_idx} OFFSET ${offset_idx}
-        """
-        rows = await db.fetch(query, *params_list)
-        history = []
-        for row in rows:
-            history.append(
-                {
-                    "id": row[0],
-                    "timestamp": row[1],
-                    "client_ip": row[2],
-                    "client_id": row[3],
-                    "user": {
-                        "id": row[4],
-                        "username": row[16],
-                        "display_name": row[17],
-                        "email": row[18],
-                    }
-                    if row[4]
-                    else None,
-                    "track": {
-                        "id": row[5],
-                        "title": row[6],
-                        "artist": row[7],
-                        "album": row[8],
-                        "art_sha1": row[19],  # New standard
-                        "duration_seconds": row[10],
-                        "codec": row[11],
-                        "bit_depth": row[12],
-                        "sample_rate_hz": row[13],
-                        "release_date": row[14],
-                        "mb_release_id": row[15],
-                    },
-                }
-            )
-        return history
-    return []
-
-
-@router.get("/api/player/history/stats")
-async def get_playback_history_stats(
-    response: Response,
-    request: Request,
-    scope: str = "all",
-    days: int = 7,
-):
-    response.headers["Cache-Control"] = "no-cache, no-store, must-revalidate"
-    response.headers["Pragma"] = "no-cache"
-    response.headers["Expires"] = "0"
-    days = max(1, min(days, 365))
-
-    async for db in get_db():
-        user_row, _ = await get_session_user(db, request.cookies.get("jamarr_session"))
-        from datetime import timedelta
-
-        where_clauses = ["DATE(timestamp) >= CURRENT_DATE + CAST($1 AS INTERVAL)"]
-        params: List[Any] = [timedelta(days=-(days - 1))]
-        if scope == "mine" and user_row:
-            where_clauses.append("h.user_id = $2")
-            params.append(user_row["id"])
-        where_sql = " AND ".join(where_clauses)
-
-        # Plays per day
-        daily_query = f"""
-            SELECT DATE(timestamp) as day, COUNT(*) as plays
-            FROM playback_history h
-            WHERE {where_sql}
-            GROUP BY day
-            ORDER BY day DESC
-        """
-        rows = await db.fetch(daily_query, *params)
-        daily = [{"day": row[0], "plays": row[1]} for row in rows]
-
-        # Top artists
-        artists_query = f"""
-            SELECT 
-                COALESCE(NULLIF(t.album_artist, ''), t.artist) as artist_name, 
-                MIN(t.artwork_id) as artwork_id, 
-                MAX(a.sha1) as art_sha1, 
-                COUNT(*) as plays
-            FROM playback_history h
-            JOIN track t ON t.id = h.track_id
-            LEFT JOIN artwork a ON t.artwork_id = a.id
-            WHERE {where_sql}
-            GROUP BY COALESCE(NULLIF(t.album_artist, ''), t.artist)
-            ORDER BY plays DESC
-            LIMIT 10
-        """
-        rows = await db.fetch(artists_query, *params)
-        artists = [
-            {
-                "artist": row[0],
-                "art_sha1": row[2],
-                "plays": row[3],
-            }
-            for row in rows
-            if row[0]
-        ]
-
-        # Top albums
-        albums_query = f"""
-            SELECT 
-                t.album, 
-                COALESCE(NULLIF(t.album_artist, ''), t.artist) as artist_name, 
-                MIN(t.artwork_id) as artwork_id, 
-                MAX(a.sha1) as art_sha1, 
-                MAX(t.release_mbid) as mb_release_id,
-                COUNT(*) as plays
-            FROM playback_history h
-            JOIN track t ON t.id = h.track_id
-            LEFT JOIN artwork a ON t.artwork_id = a.id
-                WHERE {where_sql}
-                GROUP BY t.album, COALESCE(NULLIF(t.album_artist, ''), t.artist)
-                ORDER BY plays DESC
-                LIMIT 10
-            """
-        rows = await db.fetch(albums_query, *params)
-        albums = [
-            {
-                "album": row[0],
-                "artist": row[1],
-                "art_sha1": row[3],
-                "mb_release_id": row[4],
-                "plays": row[5],
-            }
-            for row in rows
-            if row[0]
-        ]
-
-        # Top tracks
-        tracks_query = f"""
-        SELECT t.id, t.title, t.artist, t.album, t.release_mbid, t.artwork_id, MAX(a.sha1) as art_sha1, COUNT(*) as plays
-        FROM playback_history h
-        JOIN track t ON t.id = h.track_id
-        LEFT JOIN artwork a ON t.artwork_id = a.id
-        WHERE {where_sql}
-            GROUP BY t.id, t.title, t.artist, t.album, t.release_mbid, t.artwork_id
-            ORDER BY plays DESC
-            LIMIT 10
-        """
-        rows = await db.fetch(tracks_query, *params)
-        tracks = [
-            {
-                "id": row[0],
-                "title": row[1],
-                "artist": row[2],
-                "album": row[3],
-                "mb_release_id": row[4],
-                "art_sha1": row[6],
-                "plays": row[7],
-            }
-            for row in rows
-        ]
-
-        return {
-            "daily": daily,
-            "artists": artists,
-            "albums": albums,
-            "tracks": tracks,
-        }
-    return {"daily": [], "artists": [], "albums": [], "tracks": []}
 
 
 @router.get("/api/player/state", response_model=PlayerState)
