+++ app/main.py
@@ -35,7 +35,7 @@ from fastapi.staticfiles import StaticFiles  # noqa: E402
 from fastapi.responses import FileResponse  # noqa: E402
 from pathlib import Path  # noqa: E402
 from app.media import art  # noqa: E402
-from app.api import library, stream, player, search, scan, auth, media_quality, charts, lastfm  # noqa: E402
+from app.api import library, stream, player, search, scan, auth, media_quality, charts, lastfm, history  # noqa: E402
 from app import playlist  # noqa: E402
 
 app.include_router(art.router)
@@ -52,6 +52,7 @@ app.include_router(playlist.router)
 app.include_router(playlist.router)
 app.include_router(charts.router, prefix="/api")
 app.include_router(lastfm.router)
+app.include_router(history.router)
 
 
 # Serve built SvelteKit frontend (output lives in web/build)
