+++ web/src/routes/history/+page.ts
@@ -1,5 +1,6 @@
 export async function load({ fetch, url }) {
     const scope = url.searchParams.get('scope') || 'mine';
+    const source = url.searchParams.get('source') || 'all';
     const daysParam = parseInt(url.searchParams.get('days') || '7', 10);
     const days = Number.isFinite(daysParam) ? Math.max(1, Math.min(daysParam, 365)) : 7;
 
@@ -7,12 +8,12 @@ export async function load({ fetch, url }) {
     const page = Number.isFinite(pageParam) ? Math.max(1, pageParam) : 1;
 
     const [historyRes, statsRes] = await Promise.all([
-        fetch(`/api/player/history?scope=${scope}&days=${days}&page=${page}`),
-        fetch(`/api/player/history/stats?scope=${scope}&days=${days}`)
+        fetch(`/api/history/tracks?scope=${scope}&source=${source}&days=${days}&page=${page}`),
+        fetch(`/api/history/stats?scope=${scope}&source=${source}&days=${days}`)
     ]);
 
     const history = historyRes.ok ? await historyRes.json() : [];
     const stats = statsRes.ok ? await statsRes.json() : { daily: [], artists: [], albums: [], tracks: [] };
 
-    return { history, scope, days, page, stats };
+    return { history, scope, source, days, page, stats };
 }
